Improve unit tests
This change performs some minor improvements on the project's unit
tests. The idea is to have exceptions thrown up to the test methods
level, and let the testing framework deal with them. Test frameworks
already relate uncaught exceptions as failures. With this change, users
should be more aware of errors in tests, the code is simpler, and
exceptions are not _swallowed_.